NFC: Move bpforc into ast/bpforc
Previously, the link order was wrong. ast.a depended on libbpftrace.a,
AND libbpftrace.a depended on ast.a. This happened to be OK b/c
bpftrace.cpp used symbols in bpforc.cpp, thus keeping the bpforc.cpp
symbols in libbpftrace.a. Then during final linking, ast.a successfully
linked against the bpforc symbols in libbpftrace.a.

The next commit removes bpforc.cpp usage from bpftrace.cpp, breaking
linkage. This commit fixes the issue by linking ast.a directly against
bpforc. Thus, the bad link order is fixed.
