+++ src/ast/bpforc/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(bpforc bpforc.cpp)
