+++ src/ast/codegen_llvm.cpp
@@ -2,7 +2,7 @@
 #include "arch/arch.h"
 #include "ast.h"
 #include "ast/async_event_types.h"
-#include "bpforc.h"
+#include "ast/bpforc/bpforc.h"
 #include "codegen_helper.h"
 #include "log.h"
 #include "parser.tab.hh"
