+++ tests/codegen/general.cpp
@@ -1,8 +1,8 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
+#include "ast/bpforc/bpforc.h"
 #include "ast/semantic_analyser.h"
-#include "bpforc.h"
 #include "bpftrace.h"
 #include "clang_parser.h"
 #include "codegen_llvm.h"
