+++ tests/codegen/regression.cpp
@@ -1,9 +1,9 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
+#include "ast/bpforc/bpforc.h"
 #include "ast/resource_analyser.h"
 #include "ast/semantic_analyser.h"
-#include "bpforc.h"
 #include "bpftrace.h"
 #include "clang_parser.h"
 #include "codegen_llvm.h"
