+++ tests/probe.cpp
@@ -1,7 +1,7 @@
+#include "ast/bpforc/bpforc.h"
 #include "ast/field_analyser.h"
 #include "ast/resource_analyser.h"
 #include "ast/semantic_analyser.h"
-#include "bpforc.h"
 #include "bpftrace.h"
 #include "clang_parser.h"
 #include "codegen_llvm.h"
